requirements.txt
flask
flask-cache
flask-login==0.2.11
flask-markdown
pymongo
pyyaml
pyflakes
unittest2
coverage
nose
cython
markupsafe
git+https://github.com/jwbober/conrey-dirichlet-characters.git@master#egg=dirichlet_conrey
